+++ RiskOfChaos/Networking/NetworkMessageManager.cs
@@ -12,6 +12,8 @@ namespace RiskOfChaos.Networking
 
             NetworkingAPI.RegisterMessageType<TeleportBodyMessage>();
             NetworkingAPI.RegisterMessageType<SetObjectDontDestroyOnLoadMessage>();
+
+            NetworkingAPI.RegisterMessageType<PickupTransformationNotificationMessage>();
         }
     }
 }